# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Paylist API
  version: 1.0.0
  description: API documentation for authentication and payment endpoints
servers:
  - url: http://localhost:3000/api
paths:
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: mysecurepassword
              required:
                - email
                - password
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Invalid input
...